etl/etl/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
faker==33.1.0
apscheduler>=3.10.4
